requirements.txt
python-telegram-bot
groq
python-dotenv